[FIX] fix of the motif transformers
