+++ src/protrend/transform/motif/__init__.py
@@ -1,0 +1,2 @@
+from .source import SourceTransformer, SourceToMotifConnector
+from .motif import MotifTransformer, MotifToTFBSConnector, MotifToOrganismConnector, MotifToRegulatorConnector
